+++ projects/ngx-enaio/src/lib/interfaces/enaio-document-search-options.ts
@@ -57,3 +57,4 @@ export interface EnaioDocumentSearchOptions {
    */
   fieldsschema?: EnaioFieldSchema;
 }
+